test(web themes): add snapshot tests for all themes
