+++ packages/web/src/Themes/index.ts
@@ -1,19 +1,6 @@
-import {
-  themes as styleThemes,
-} from "@naturacosmeticos/natds-styles";
-import { parseTheme } from "./parseTheme";
-import { ThemeBrandName } from "./ThemeBrandName";
-import createThemesObject from "./createThemesObject";
-import { IThemesWeb } from "./IThemesWeb";
+import { themes } from "./themes";
 
 export { IThemeWeb } from "./IThemeWeb";
+export { themes } from "./themes";
 
-export const themes = Object.keys(styleThemes).reduce((value: IThemesWeb<ThemeBrandName>, key: string) => {
-
-  const result = value;
-
-  result[key as ThemeBrandName].dark = parseTheme(styleThemes[key as ThemeBrandName].dark);
-  result[key as ThemeBrandName].light = parseTheme(styleThemes[key as ThemeBrandName].light);
-
-  return result;
-}, createThemesObject());
+export default themes;
